Crate actions_toolkit_sys

Raw bindings to the GitHub actions toolkit API for projects using wasm-bindgen.

Modules

core
Core functions for setting results, logging, registering secrets and exporting variables.

exec
Execute your tools on the command line in a cross-platform way.

github
An Octokit client hydrated with the context of the current action.

io
Core functions for CLI filesystem scenarios.

tool_cache
Functions necessary for downloading and caching tools.