Module actions_toolkit_sys::core

Core functions for setting results, logging, registering secrets and exporting variables.

Structs

InputOptions
Interface for getInput options

Enums

ExitCode
The code to exit an action

Functions

add_path
Prepends inputPath to the PATH (for this action and future actions).

debug
Writes debug message to user log.

end_group
End an output group.

error
Adds an error issue.

export_variable
Sets env variable for this action and future actions in the job.

get_input
Gets the value of an input. The value is also trimmed.

get_state

group
Wrap an asynchronous function call in a group.

info
Writes info to log with console.log.

issue

issue_command

save_state

set_failed
Sets the action status to failed. When the action exits it will be with an exit code of 1.

set_output
Sets the value of an output.

set_secret
Registers a secret which will get masked from logs.

start_group
Begin an output group.

warning
Adds an warning issue.