Module actions_toolkit_sys::io

Core functions for CLI filesystem scenarios.

Structs

CopyOptions
Interface for cp options.

MoveOptions
Interface for mv options.

Functions

cp
Copies a file or folder.

mkdir_p
Make a directory. Creates the full path with folders in between. Will throw if it fails.

mv
Moves a path.

rm_rf
Remove a path recursively with force.

which
Returns path of a tool had the tool actually been invoked. Resolves via paths. If you check and the tool does not exist, it will throw.